requirements.txt
sstreamlit
yfinance
GoogleNews
plotly
pandas
google-generativeai